frontend/requirements.txt
discord.py>=1.7.2
python-dotenv
py4j